bump to 20.03
Update Codeblocks to the 20.03 release and also bump the runtime version to 20.08.
The package seems to be abandoned by the maintainer and the current Freedesktop 1.6 runtime is terribly outdated.

This was not thoroughly tested, it would be nice to some feedback from users, so please test the CI test build.

Closes #3 and #5.

Notable changes:

* Switch the default terminal from the ancient Xterm to Termite, a Wayland compatible VTE terminal.
Maybe reconsider switching to something else in the future.
I'm using a sort of a hack, include a single double-quote to get Codeblock to quote the whole command string given to Termite.
* More access permissions to the host. This is a development package.
* Added app icon from https://www.codeblocks.org/images/logo160.png, resized to 128x128.