enable contrib plugins
Package also the contrib plugins.
Boost module was added as it's a dependecy of NassiShneiderman plugin.
FileManager was disabled as it depends on python2 gamin lib.
dragscroll was disabled as it's failing to be compiled on aarch64.
See list of plugins at http://wiki.codeblocks.org/index.php/Code::Blocks_Plugins#Contrib_Plugins

Closes #4.